requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
pdfplumber==0.10.4
python-docx==1.1.0
chromadb==0.4.22
scikit-learn==1.4.0
numpy==1.26.3
python-multipart==0.0.6

# Testing dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
python-dotenv==1.0.0

# Demo dependencies
reportlab==4.0.9

